Refactor studentID types from int back to Long

+++ curriculum-service/src/main/java/com/project/curriculumservice/controller/CourseOfferingController.java
@@ -38,7 +38,7 @@ public class CourseOfferingController {
     }
 
     @GetMapping("/student/{studentId}")
-    public ResponseEntity<List<CourseOfferingResponseDTO>>  getCourseOfferingsForStudent(@PathVariable int studentId) {
+    public ResponseEntity<List<CourseOfferingResponseDTO>>  getCourseOfferingsForStudent(@PathVariable Long studentId) {
         return ResponseEntity.ok(
                 courseOfferingService.getCourseOfferingsForStudent(studentId)
         );
